get ready for imgui

+++ Seagull Core/src/Input/Input.h
@@ -11,9 +11,9 @@ namespace SG
 		static bool IsKeyPressed(int keycode);
 		static bool IsMouseButtonPressed(int button);
 
-		static std::pair<LONG, LONG> GetMousePos();
-		static LONG GetMouseX();
-		static LONG GetMouseY();
+		static std::pair<float, float> GetMousePos();
+		static float GetMouseX();
+		static float GetMouseY();
 	};
 
 }
